update main script and readme
