+++ .vscode/launch.json
@@ -6,10 +6,13 @@
   "configurations": [
     {
       "name": "Python: FastAPI",
-      "type": "python",
+      "type": "debugpy",
       "request": "launch",
       "module": "uvicorn",
-      "args": ["app:app", "--reload"],
+      "args": [
+        "app.main:app",
+        "--reload"
+      ],
       "jinja": true,
       "justMyCode": true
     }
