+++ start.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-gunicorn --worker-class uvicorn.workers.UvicornWorker -w 4 -b :8000 "app:create_app()"
+gunicorn --worker-class uvicorn.workers.UvicornWorker -w 4 -b :8000 "app.main:app"
